+++ packages/component/src/ActivityFeedback/private/FeedbackForm.tsx
@@ -1,49 +1,28 @@
 import { hooks } from 'botframework-webchat-api';
-import classNames from 'classnames';
 import React, { memo, useCallback, useEffect, useRef, useState, type FormEventHandler } from 'react';
-import { useRefFrom } from 'use-ref-from';
-import { useStateWithRef } from 'use-state-with-ref';
 
 import Markdownable from '../../Attachment/Text/private/Markdownable';
-import useStyleSet from '../../hooks/useStyleSet';
 import testIds from '../../testIds';
-import useActivity from '../providers/useActivity';
-import useSelectedAction from '../providers/useSelectedAction';
-import useSubmitCallback from '../providers/useSubmitCallback';
+import useActivityFeedbackHooks from '../providers/useActivityFeedbackHooks';
 import FeedbackTextArea from './FeedbackTextArea';
-import getDisclaimer from './getDisclaimer';
+import getDisclaimerFromReviewAction from './getDisclaimerFromReviewAction';
 
 const { useLocalizer } = hooks;
 
 function FeedbackForm() {
-  const [{ feedbackForm }] = useStyleSet();
-  const [activity] = useActivity();
+  const { useFeedbackText, useSelectedAction } = useActivityFeedbackHooks();
+
+  const [selectedAction] = useSelectedAction();
   const [hasFocus, setHasFocus] = useState(false);
-  const [selectedAction, setSelectedAction] = useSelectedAction();
-  const [userFeedback, setUserFeedback, userFeedbackRef] = useStateWithRef('');
+  const [userFeedback, setUserFeedback] = useFeedbackText();
   const feedbackTextAreaRef = useRef<HTMLTextAreaElement>(null);
   const localize = useLocalizer();
-  const submit = useSubmitCallback();
 
-  const disclaimer = getDisclaimer(activity);
-  const selectedActionRef = useRefFrom(selectedAction);
-
-  const handleCancelButtonClick = useCallback(() => {
-    setSelectedAction(undefined);
-  }, [setSelectedAction]);
+  const disclaimer = getDisclaimerFromReviewAction(selectedAction);
 
   const handleMessageChange: FormEventHandler<HTMLTextAreaElement> = useCallback(
     ({ currentTarget: { value } }) => setUserFeedback(value),
     [setUserFeedback]
-  );
-
-  const handleSubmit = useCallback<FormEventHandler<HTMLFormElement>>(
-    event => {
-      event.preventDefault();
-
-      submit(selectedActionRef.current, userFeedbackRef.current);
-    },
-    [selectedActionRef, submit, userFeedbackRef]
   );
 
   useEffect(() => {
@@ -58,8 +37,8 @@ function FeedbackForm() {
   }, [feedbackTextAreaRef, hasFocus, setHasFocus]);
 
   return (
-    <form className={classNames('webchat__feedback-form', feedbackForm + '')} onSubmit={handleSubmit}>
-      <span className={classNames('webchat__feedback-form__body')}>{localize('FEEDBACK_FORM_TITLE')}</span>
+    <div className="webchat__feedback-form__form">
+      <span className="webchat__feedback-form__form-header">{localize('FEEDBACK_FORM_TITLE')}</span>
       <FeedbackTextArea
         data-testid={testIds.feedbackSendBox}
         onInput={handleMessageChange}
@@ -68,20 +47,16 @@ function FeedbackForm() {
         startRows={3}
         value={userFeedback}
       />
-      {disclaimer && <Markdownable className={classNames('webchat__feedback-form__caption')} text={disclaimer} />}
-      <div className={classNames('webchat__feedback-form__container')}>
-        <button className={classNames('webchat__feedback-form__submit-button')} type="submit">
+      {disclaimer && <Markdownable className="webchat__feedback-form__form-footer" text={disclaimer} />}
+      <div className="webchat__feedback-form__submission-button-bar">
+        <button className="webchat__feedback-form__submit-button" type="submit">
           {localize('FEEDBACK_FORM_SUBMIT_BUTTON_LABEL')}
         </button>
-        <button
-          className={classNames('webchat__feedback-form__cancel-button')}
-          onClick={handleCancelButtonClick}
-          type="button"
-        >
+        <button className="webchat__feedback-form__cancel-button" type="reset">
           {localize('FEEDBACK_FORM_CANCEL_BUTTON_LABEL')}
         </button>
       </div>
-    </form>
+    </div>
   );
 }