+++ packages/component/src/ActivityFeedback/private/FeedbackVoteButtonBar.tsx
@@ -1,25 +1,34 @@
-import React, { Fragment, memo, type ReactNode } from 'react';
+import { validateProps } from 'botframework-webchat-api/internal';
+import React, { memo } from 'react';
+import { literal, object, pipe, readonly, union, type InferInput } from 'valibot';
 
-import useActions from '../providers/useActions';
+import useActivityFeedbackHooks from '../providers/useActivityFeedbackHooks';
 import FeedbackVoteButton from './FeedbackVoteButton';
 
-type FeedbackVoteButtonBarProps = Readonly<{
-  children?: ReactNode | undefined;
-}>;
+const feedbackVoteButtonBarPropsSchema = pipe(
+  object({
+    buttonAs: union([literal('button'), literal('radio')])
+  }),
+  readonly()
+);
 
-function FeedbackVoteButtonBar() {
+type FeedbackVoteButtonBarProps = InferInput<typeof feedbackVoteButtonBarPropsSchema>;
+
+function FeedbackVoteButtonBar(props: FeedbackVoteButtonBarProps) {
+  const { buttonAs } = validateProps(feedbackVoteButtonBarPropsSchema, props);
+
+  const { useActions } = useActivityFeedbackHooks();
+
   const [actions] = useActions();
 
   return (
-    <Fragment>
+    <div className="webchat__feedback-form__vote-button-bar">
       {actions.map((action, index) => (
-        <FeedbackVoteButton action={action} key={action['@id'] || index} />
+        <FeedbackVoteButton action={action} as={buttonAs} key={action['@id'] || index} />
       ))}
-    </Fragment>
+    </div>
   );
 }
 
-FeedbackVoteButtonBar.displayName = 'FeedbackVoteButtonBar';
-
 export default memo(FeedbackVoteButtonBar);
-export { type FeedbackVoteButtonBarProps };
+export { feedbackVoteButtonBarPropsSchema, type FeedbackVoteButtonBarProps };