+++ packages/component/src/ActivityFeedback/private/getDisclaimer.ts
@@ -1,6 +1,0 @@
-import { type WebChatActivity } from 'botframework-webchat-core';
-import { hasDisclaimer } from './hasFeedbackLoop';
-
-export default function getDisclaimer(activity: WebChatActivity): string | undefined {
-  return hasDisclaimer(activity) ? activity.channelData.feedbackLoop.disclaimer : undefined;
-}