+++ packages/component/src/ActivityFeedback/providers/private/useActivityFeedbackContext.ts
@@ -1,6 +1,0 @@
-import { useContext } from 'react';
-import ActivityFeedbackContext, { type ActivityFeedbackContextType } from './ActivityFeedbackContext';
-
-export default function useActivityFeedbackContext(): ActivityFeedbackContextType {
-  return useContext(ActivityFeedbackContext);
-}