+++ packages/component/src/ActivityFeedback/providers/useShouldAllowResubmit.ts
@@ -1,5 +1,0 @@
-import useActivityFeedbackContext from './private/useActivityFeedbackContext';
-
-export default function useShouldAllowResubmit(): readonly [boolean] {
-  return useActivityFeedbackContext().shouldAllowResubmitState;
-}