+++ packages/component/src/testIds.ts
@@ -5,6 +5,7 @@ const testIds = {
   sendBoxTextBox: 'send box text area',
   typingIndicator: 'typing indicator',
   viewCodeButton: 'view code button',
+  feedbackButton: 'feedback button',
   feedbackSendBox: 'feedback sendbox'
 };